requirements.txt
click~=8.1.4
setuptools~=65.5.1
boto3~=1.28.2
botocore~=1.31.2
python-gnupg~=0.5.0
python-dotenv~=1.0.0